fix: Propagate UnconvertibleExpression through OR
If either half of an OR is an unconvertible expression,
we reduce the entire filter to an ALWAYS_TRUE, which means
that no filtering is pushed down.

This was discovered by the TPC-DS query

```
select * from customer_address
where
  (ca_country = 'United States' AND ca_state IN( 'SC', 'IN', 'VA'))
  OR
  (ca_country = 'United States' AND ca_state IN( 'WA', 'KS', 'KY'))
```

The IN expression is not currently pushable, but we can fix that by
implementing it in Vortex.